fix: cast statvfs values to u64 for portable disk-space preflight
The nix crate's statvfs returns platform-dependent integer types for
blocks_available() and fragment_size(). On some targets these are u32,
causing type mismatches with the u64 check_free_space signature. Cast
both to u64 (safe widening) before the checked_mul arithmetic.

+++ src-tauri/src/infrastructure/sqlite/migrations/v20260730025727_chat_message_blocks_thinking_kind.rs
@@ -176,9 +176,8 @@ fn ensure_sufficient_free_space(conn: &Connection) -> AppResult<()> {
         }
     };
     let available_bytes =
-        match filesystem
-            .blocks_available()
-            .checked_mul(filesystem.fragment_size())
+        match (filesystem.blocks_available() as u64)
+            .checked_mul(filesystem.fragment_size() as u64)
         {
             Some(bytes) => bytes,
             None => {
